+++ web/documentserver-example/java/src/main/java/entities/FileModel.java
@@ -149,21 +149,26 @@ public class FileModel {
 
     // change the document type
     public void changeType(final String modeParam, final String typeParam, final User user, final String fileName) {
+        // check if the file with such an extension can be edited
+        String fileExt = FileUtility.getFileExtension(document.getTitle());
+        Boolean canFill = DocumentManager.getFillExts().contains(fileExt);
+        Boolean canEdit = DocumentManager.getEditedExts().contains(fileExt);
+
         if (modeParam != null) {
             mode = modeParam;
+        } else {
+            mode = canFill ? "fillForms" : "edit";
         }
         if (typeParam != null) {
             type = typeParam;
         }
 
-        // check if the file with such an extension can be edited
-        String fileExt = FileUtility.getFileExtension(document.getTitle());
-        Boolean canEdit = DocumentManager.getEditedExts().contains(fileExt);
         // check if the Submit form button is displayed or not
-        editorConfig.getCustomization().setSubmitForm(true);
+        if (mode.equals("fillForms") || mode.equals("embedded")) {
+            editorConfig.getCustomization().setSubmitForm(user.getId().equals("uid-1"));
+        }
 
-        if ((!canEdit && mode.equals("edit") || mode.equals("fillForms"))
-                && DocumentManager.getFillExts().contains(fileExt)) {
+        if ((!canEdit && mode.equals("edit") || mode.equals("fillForms")) && canFill) {
             canEdit = true;
             mode = "fillForms";
         }