+++ web/documentserver-example/python/src/configuration/configuration.py
@@ -22,7 +22,7 @@ from src.common import string
 
 
 class ConfigurationManager:
-    version = '1.9.0'
+    version = '1.10.0'
 
     def getVersion(self) -> str:
         return self.version