+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+- csharp: filling by default
+- csharp-mvc: filling by default
+- java: filling by default
+- java-spring: filling by default
+- php: filling by default
+- python: filling by default
+- ruby: filling by default
+- csharp: creating and editing pdf instead docxf
+- csharp-mvc: creating and editing pdf instead docxf
+- java: creating and editing pdf instead docxf
+- java-spring: creating and editing pdf instead docxf
+- php: creating and editing pdf instead docxf
+- python: creating and editing pdf instead docxf
+- ruby: creating and editing pdf instead docxf
 - nodejs: tabs menu
 - nodejs: converting function on index page
 - java-spring: using java docs-integration-sdk